+++ PseudoGameEngine/math/Vector2.cs
@@ -44,8 +44,8 @@ namespace PseudoGameEngine.math
 
         #region Public Fields
 
-        public double X;
-        public double Y;
+        public float X;
+        public float Y;
 
         #endregion Public Fields
 
@@ -77,13 +77,13 @@ namespace PseudoGameEngine.math
 
         #region Constructors
 
-        public Vector2(double x, double y)
+        public Vector2(float x, float y)
         {
             this.X = x;
             this.Y = y;
         }
 
-        public Vector2(double value)
+        public Vector2(float value)
         {
             this.X = value;
             this.Y = value;
@@ -107,69 +107,69 @@ namespace PseudoGameEngine.math
             result.Y = value1.Y + value2.Y;
         }
 
-        public static Vector2 Barycentric(Vector2 value1, Vector2 value2, Vector2 value3, double amount1, double amount2)
+        public static Vector2 Barycentric(Vector2 value1, Vector2 value2, Vector2 value3, float amount1, float amount2)
         {
             return new Vector2(
-                MathHelper.Barycentric(value1.X, value2.X, value3.X, amount1, amount2),
-                MathHelper.Barycentric(value1.Y, value2.Y, value3.Y, amount1, amount2));
+                (float)MathHelper.Barycentric(value1.X, value2.X, value3.X, amount1, amount2),
+                (float)MathHelper.Barycentric(value1.Y, value2.Y, value3.Y, amount1, amount2));
         }
 
-        public static void Barycentric(ref Vector2 value1, ref Vector2 value2, ref Vector2 value3, double amount1, double amount2, out Vector2 result)
+        public static void Barycentric(ref Vector2 value1, ref Vector2 value2, ref Vector2 value3, float amount1, float amount2, out Vector2 result)
         {
             result = new Vector2(
-                MathHelper.Barycentric(value1.X, value2.X, value3.X, amount1, amount2),
-                MathHelper.Barycentric(value1.Y, value2.Y, value3.Y, amount1, amount2));
+                (float)MathHelper.Barycentric(value1.X, value2.X, value3.X, amount1, amount2),
+                (float)MathHelper.Barycentric(value1.Y, value2.Y, value3.Y, amount1, amount2));
         }
 
-        public static Vector2 CatmullRom(Vector2 value1, Vector2 value2, Vector2 value3, Vector2 value4, double amount)
+        public static Vector2 CatmullRom(Vector2 value1, Vector2 value2, Vector2 value3, Vector2 value4, float amount)
         {
             return new Vector2(
-                MathHelper.CatmullRom(value1.X, value2.X, value3.X, value4.X, amount),
-                MathHelper.CatmullRom(value1.Y, value2.Y, value3.Y, value4.Y, amount));
+                (float)MathHelper.CatmullRom(value1.X, value2.X, value3.X, value4.X, amount),
+                (float)MathHelper.CatmullRom(value1.Y, value2.Y, value3.Y, value4.Y, amount));
         }
 
-        public static void CatmullRom(ref Vector2 value1, ref Vector2 value2, ref Vector2 value3, ref Vector2 value4, double amount, out Vector2 result)
+        public static void CatmullRom(ref Vector2 value1, ref Vector2 value2, ref Vector2 value3, ref Vector2 value4, float amount, out Vector2 result)
         {
             result = new Vector2(
-                MathHelper.CatmullRom(value1.X, value2.X, value3.X, value4.X, amount),
-                MathHelper.CatmullRom(value1.Y, value2.Y, value3.Y, value4.Y, amount));
+                (float)MathHelper.CatmullRom(value1.X, value2.X, value3.X, value4.X, amount),
+                (float)MathHelper.CatmullRom(value1.Y, value2.Y, value3.Y, value4.Y, amount));
         }
 
         public static Vector2 Clamp(Vector2 value1, Vector2 min, Vector2 max)
         {
             return new Vector2(
-                MathHelper.Clamp(value1.X, min.X, max.X),
-                MathHelper.Clamp(value1.Y, min.Y, max.Y));
+                (float)MathHelper.Clamp(value1.X, min.X, max.X),
+                (float)MathHelper.Clamp(value1.Y, min.Y, max.Y));
         }
 
         public static void Clamp(ref Vector2 value1, ref Vector2 min, ref Vector2 max, out Vector2 result)
         {
             result = new Vector2(
-                MathHelper.Clamp(value1.X, min.X, max.X),
-                MathHelper.Clamp(value1.Y, min.Y, max.Y));
+                (float)MathHelper.Clamp(value1.X, min.X, max.X),
+                (float)MathHelper.Clamp(value1.Y, min.Y, max.Y));
         }
 
-        public static double Distance(Vector2 value1, Vector2 value2)
+        public static float Distance(Vector2 value1, Vector2 value2)
         {
-            double v1 = value1.X - value2.X, v2 = value1.Y - value2.Y;
-            return (double)Math.Sqrt((v1 * v1) + (v2 * v2));
+            float v1 = value1.X - value2.X, v2 = value1.Y - value2.Y;
+            return (float)Math.Sqrt((v1 * v1) + (v2 * v2));
         }
 
-        public static void Distance(ref Vector2 value1, ref Vector2 value2, out double result)
+        public static void Distance(ref Vector2 value1, ref Vector2 value2, out float result)
         {
-            double v1 = value1.X - value2.X, v2 = value1.Y - value2.Y;
-            result = (double)Math.Sqrt((v1 * v1) + (v2 * v2));
+            float v1 = value1.X - value2.X, v2 = value1.Y - value2.Y;
+            result = (float)Math.Sqrt((v1 * v1) + (v2 * v2));
         }
 
-        public static double DistanceSquared(Vector2 value1, Vector2 value2)
+        public static float DistanceSquared(Vector2 value1, Vector2 value2)
         {
-            double v1 = value1.X - value2.X, v2 = value1.Y - value2.Y;
+            float v1 = value1.X - value2.X, v2 = value1.Y - value2.Y;
             return (v1 * v1) + (v2 * v2);
         }
 
-        public static void DistanceSquared(ref Vector2 value1, ref Vector2 value2, out double result)
+        public static void DistanceSquared(ref Vector2 value1, ref Vector2 value2, out float result)
         {
-            double v1 = value1.X - value2.X, v2 = value1.Y - value2.Y;
+            float v1 = value1.X - value2.X, v2 = value1.Y - value2.Y;
             result = (v1 * v1) + (v2 * v2);
         }
 
@@ -186,27 +186,27 @@ namespace PseudoGameEngine.math
             result.Y = value1.Y / value2.Y;
         }
 
-        public static Vector2 Divide(Vector2 value1, double divider)
+        public static Vector2 Divide(Vector2 value1, float divider)
         {
-            double factor = 1 / divider;
+            float factor = 1 / divider;
             value1.X *= factor;
             value1.Y *= factor;
             return value1;
         }
 
-        public static void Divide(ref Vector2 value1, double divider, out Vector2 result)
+        public static void Divide(ref Vector2 value1, float divider, out Vector2 result)
         {
-            double factor = 1 / divider;
+            float factor = 1 / divider;
             result.X = value1.X * factor;
             result.Y = value1.Y * factor;
         }
 
-        public static double Dot(Vector2 value1, Vector2 value2)
+        public static float Dot(Vector2 value1, Vector2 value2)
         {
             return (value1.X * value2.X) + (value1.Y * value2.Y);
         }
 
-        public static void Dot(ref Vector2 value1, ref Vector2 value2, out double result)
+        public static void Dot(ref Vector2 value1, ref Vector2 value2, out float result)
         {
             result = (value1.X * value2.X) + (value1.Y * value2.Y);
         }
@@ -229,7 +229,7 @@ namespace PseudoGameEngine.math
         public static Vector2 Reflect(Vector2 vector, Vector2 normal)
         {
             Vector2 result;
-            double val = 2.0f * ((vector.X * normal.X) + (vector.Y * normal.Y));
+            float val = 2.0f * ((vector.X * normal.X) + (vector.Y * normal.Y));
             result.X = vector.X - (normal.X * val);
             result.Y = vector.Y - (normal.Y * val);
             return result;
@@ -237,7 +237,7 @@ namespace PseudoGameEngine.math
 
         public static void Reflect(ref Vector2 vector, ref Vector2 normal, out Vector2 result)
         {
-            double val = 2.0f * ((vector.X * normal.X) + (vector.Y * normal.Y));
+            float val = 2.0f * ((vector.X * normal.X) + (vector.Y * normal.Y));
             result.X = vector.X - (normal.X * val);
             result.Y = vector.Y - (normal.Y * val);
         }
@@ -247,41 +247,41 @@ namespace PseudoGameEngine.math
             return X.GetHashCode() + Y.GetHashCode();
         }
 
-        public static Vector2 Hermite(Vector2 value1, Vector2 tangent1, Vector2 value2, Vector2 tangent2, double amount)
+        public static Vector2 Hermite(Vector2 value1, Vector2 tangent1, Vector2 value2, Vector2 tangent2, float amount)
         {
             Vector2 result = new Vector2();
             Hermite(ref value1, ref tangent1, ref value2, ref tangent2, amount, out result);
             return result;
         }
 
-        public static void Hermite(ref Vector2 value1, ref Vector2 tangent1, ref Vector2 value2, ref Vector2 tangent2, double amount, out Vector2 result)
+        public static void Hermite(ref Vector2 value1, ref Vector2 tangent1, ref Vector2 value2, ref Vector2 tangent2, float amount, out Vector2 result)
         {
-            result.X = MathHelper.Hermite(value1.X, tangent1.X, value2.X, tangent2.X, amount);
-            result.Y = MathHelper.Hermite(value1.Y, tangent1.Y, value2.Y, tangent2.Y, amount);
+            result.X = (float)MathHelper.Hermite(value1.X, tangent1.X, value2.X, tangent2.X, amount);
+            result.Y = (float)MathHelper.Hermite(value1.Y, tangent1.Y, value2.Y, tangent2.Y, amount);
         }
 
-        public double Length()
+        public float Length()
         {
-            return (double)Math.Sqrt((X * X) + (Y * Y));
+            return (float)Math.Sqrt((X * X) + (Y * Y));
         }
 
-        public double LengthSquared()
+        public float LengthSquared()
         {
             return (X * X) + (Y * Y);
         }
 
-        public static Vector2 Lerp(Vector2 value1, Vector2 value2, double amount)
+        public static Vector2 Lerp(Vector2 value1, Vector2 value2, float amount)
         {
             return new Vector2(
-                MathHelper.Lerp(value1.X, value2.X, amount),
-                MathHelper.Lerp(value1.Y, value2.Y, amount));
+                (float)MathHelper.Lerp(value1.X, value2.X, amount),
+                (float)MathHelper.Lerp(value1.Y, value2.Y, amount));
         }
 
-        public static void Lerp(ref Vector2 value1, ref Vector2 value2, double amount, out Vector2 result)
+        public static void Lerp(ref Vector2 value1, ref Vector2 value2, float amount, out Vector2 result)
         {
             result = new Vector2(
-                MathHelper.Lerp(value1.X, value2.X, amount),
-                MathHelper.Lerp(value1.Y, value2.Y, amount));
+                (float)MathHelper.Lerp(value1.X, value2.X, amount),
+                (float)MathHelper.Lerp(value1.Y, value2.Y, amount));
         }
 
         public static Vector2 Max(Vector2 value1, Vector2 value2)
@@ -315,14 +315,14 @@ namespace PseudoGameEngine.math
             return value1;
         }
 
-        public static Vector2 Multiply(Vector2 value1, double scaleFactor)
+        public static Vector2 Multiply(Vector2 value1, float scaleFactor)
         {
             value1.X *= scaleFactor;
             value1.Y *= scaleFactor;
             return value1;
         }
 
-        public static void Multiply(ref Vector2 value1, double scaleFactor, out Vector2 result)
+        public static void Multiply(ref Vector2 value1, float scaleFactor, out Vector2 result)
         {
             result.X = value1.X * scaleFactor;
             result.Y = value1.Y * scaleFactor;
@@ -349,14 +349,14 @@ namespace PseudoGameEngine.math
 
         public void Normalize()
         {
-            double val = 1.0f / (double)Math.Sqrt((X * X) + (Y * Y));
+            float val = 1.0f / (float)Math.Sqrt((X * X) + (Y * Y));
             X *= val;
             Y *= val;
         }
 
         public static Vector2 Normalize(Vector2 value)
         {
-            double val = 1.0f / (double)Math.Sqrt((value.X * value.X) + (value.Y * value.Y));
+            float val = 1.0f / (float)Math.Sqrt((value.X * value.X) + (value.Y * value.Y));
             value.X *= val;
             value.Y *= val;
             return value;
@@ -364,23 +364,23 @@ namespace PseudoGameEngine.math
 
         public static void Normalize(ref Vector2 value, out Vector2 result)
         {
-            double val = 1.0f / (double)Math.Sqrt((value.X * value.X) + (value.Y * value.Y));
+            float val = 1.0f / (float)Math.Sqrt((value.X * value.X) + (value.Y * value.Y));
             result.X = value.X * val;
             result.Y = value.Y * val;
         }
 
-        public static Vector2 SmoothStep(Vector2 value1, Vector2 value2, double amount)
+        public static Vector2 SmoothStep(Vector2 value1, Vector2 value2, float amount)
         {
             return new Vector2(
-                MathHelper.SmoothStep(value1.X, value2.X, amount),
-                MathHelper.SmoothStep(value1.Y, value2.Y, amount));
+                (float)MathHelper.SmoothStep(value1.X, value2.X, amount),
+                (float)MathHelper.SmoothStep(value1.Y, value2.Y, amount));
         }
 
-        public static void SmoothStep(ref Vector2 value1, ref Vector2 value2, double amount, out Vector2 result)
+        public static void SmoothStep(ref Vector2 value1, ref Vector2 value2, float amount, out Vector2 result)
         {
             result = new Vector2(
-                MathHelper.SmoothStep(value1.X, value2.X, amount),
-                MathHelper.SmoothStep(value1.Y, value2.Y, amount));
+                (float)MathHelper.SmoothStep(value1.X, value2.X, amount),
+                (float)MathHelper.SmoothStep(value1.Y, value2.Y, amount));
         }
 
         public static Vector2 Subtract(Vector2 value1, Vector2 value2)
@@ -421,7 +421,7 @@ namespace PseudoGameEngine.math
             Quaternion.Multiply(ref quat, ref v, out t);
             Quaternion.Multiply(ref t, ref i, out v);
 
-            result = new Vector2(v.X, v.Y);
+            result = new Vector2((float)v.X, (float)v.Y);
         }
 
         public static void Transform(
@@ -519,7 +519,7 @@ namespace PseudoGameEngine.math
         }
 
 
-        public static Vector2 operator *(Vector2 value, double scaleFactor)
+        public static Vector2 operator *(Vector2 value, float scaleFactor)
         {
             value.X *= scaleFactor;
             value.Y *= scaleFactor;
@@ -527,7 +527,7 @@ namespace PseudoGameEngine.math
         }
 
 
-        public static Vector2 operator *(double scaleFactor, Vector2 value)
+        public static Vector2 operator *(float scaleFactor, Vector2 value)
         {
             value.X *= scaleFactor;
             value.Y *= scaleFactor;
@@ -543,9 +543,9 @@ namespace PseudoGameEngine.math
         }
 
 
-        public static Vector2 operator /(Vector2 value1, double divider)
+        public static Vector2 operator /(Vector2 value1, float divider)
         {
-            double factor = 1 / divider;
+            float factor = 1 / divider;
             value1.X *= factor;
             value1.Y *= factor;
             return value1;
